Move the 'Install Package' operation to a common location.
